% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ACF.lmerMod}
\alias{ACF.lmerMod}
\title{ACF function for merMod objects}
\usage{
ACF.lmerMod(
  object,
  maxLag,
  resType = c("pearson", "response", "normalized"),
  ...
)
}
\description{
ACF function for merMod objects
}
