% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as_replaced_factor}
\alias{as_replaced_factor}
\title{Replace character elements and convert to factor.}
\usage{
as_replaced_factor(x, replacement)
}
\arguments{
\item{x}{A character vector to replace.}

\item{replacement}{A named character vector with the original character elements as
elements and desired replacement elements as names.}
}
\description{
Replace the elements in a character and convert to a factor. The levels of the factor
are the same order as the replacement vector.
}
