% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bootstrap}
\alias{bootstrap}
\title{Generalized bootstrapper}
\usage{
bootstrap(x, y = NULL, fun, boot.reps = 1000, alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{NULL (default) or a numeric vector for statistics that require a second
numeric vector}

\item{fun}{A character vector of the desired statistic function. For instance,
to calculate the variance of values in the numeric vector, pass "var".}

\item{boot.reps}{The number of bootstrapping replications.}

\item{alpha}{The significance level for calculating a confidence interval.}
}
\description{
Computes bootstrap resamples of a statistic
}
