% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfind.R
\name{cfind}
\alias{cfind}
\title{Find objects by name and class}
\usage{
cfind(what, ignore.case = TRUE, class = "any")
}
\arguments{
\item{what}{See \code{\link[utils]{apropos}}}

\item{ignore.case}{See \code{\link[utils]{apropos}}}

\item{class}{character string with a class. Objects of class \code{class} will
be returned.}
}
\description{
Wrapper of the \code{\link[utils]{apropos}} function for finding objects, but
includes the ability to find objects based on class. "c"find for "class" find.
}
