% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicore.R
\name{coreApply}
\alias{coreApply}
\title{Apply a function (optionally in parallel) to a list}
\usage{
coreApply(X, FUN, check.cores = FALSE)
}
\arguments{
\item{X}{A list of data.frames over which to apply a function. Must inherit class
\code{core.list} from the \code{\link{assign_cores}} function.}

\item{FUN}{A function.}

\item{check.cores}{Logical. Should the number of requested cores (determined as the length
of list X) be compared against the number of available cores?}
}
\description{
Apply a function over data.frames assigned to a core. If on a Unix system (Linux or Mac),
the function will be applied in parallel. If using a Windows system, the function
will not be applied in parallel
}
