% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{crossv_loo_grouped}
\alias{crossv_loo_grouped}
\title{Generate leave-one-out resamples using a grouped \code{data.frame}}
\usage{
crossv_loo_grouped(data, id = ".id")
}
\arguments{
\item{data}{A grouped data frame.}

\item{id}{Name of variable that gives each training set a unique integer id.}
}
\description{
Expands the capability of the \code{\link[modelr]{crossv_loo}} function 
such that groups in a grouped data frame form the basis of folds to be
"left out" of a model training set for cross-validation.
}
\examples{
# Normal crossv_loo
cv_iris <- modelr::crossv_loo(iris)

# Note that the number of training paritions is equal to nrow(iris)

# Use the new function
cv_iris2 <- crossv_loo_grouped(group_by(iris, Species))

# Note that the number of training partitions is equal to n_groups(data)



}
