% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{extreme_n}
\alias{extreme_n}
\title{Find the n extreme values from a data.frame}
\usage{
extreme_n(x, n, col)
}
\arguments{
\item{x}{A data frame}

\item{n}{The number of extreme value}

\item{col}{The name of the column containing the values}
}
\description{
Find the n extreme values from a data.frame
}
