% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fct_contr_sum}
\alias{fct_contr_sum}
\title{Set a factor with contrasts that sum-to-zero}
\usage{
fct_contr_sum(x, drop.levels = FALSE)
}
\arguments{
\item{x}{A factor.}

\item{drop.levels}{Logical. Should unused levels be dropped?}
}
\description{
A wrapper of \code{\link[stats]{contr.sum}}, but the resulting contrasts matrix included column names
and is easier for extracting coefficients from a fitted model object
}
