% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{find_dir}
\alias{find_dir}
\title{Search for a directory in a path}
\usage{
find_dir(pattern)
}
\arguments{
\item{pattern}{A character pattern of the directory to search for}
}
\description{
Find the full path of a directory in the path of the current working directory
}
