% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{format_numbers}
\alias{format_numbers}
\title{Format numbers for printing}
\usage{
format_numbers(x, signif.digits = 3L)
}
\arguments{
\item{x}{A numeric vector.}

\item{signif.digits}{An integer specifying the number of significant digits}
}
\description{
Formats numbers using specified significant digits for printing tables or graphs
}
