% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{labdocument}
\alias{labdocument}
\alias{templ_labdocument}
\title{Render a custom lab document from the Neyhart template}
\usage{
labdocument(...)

templ_labdocument()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for laboratory documents, mostly for create the "Lab Philosophy"
document.
}
