% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Set up a new project}
\usage{
new_project(project.name, dir = ".", use.git = TRUE, use.msi = TRUE)
}
\arguments{
\item{project.name}{The name of the project.}

\item{dir}{The root directory of the new project.}

\item{use.msi}{Logical: should the project be setup for use on a remote server?}

\item{use.this}{Logical: should a git repository be initialized?}
}
\description{
This function uses the \code{\link[usethis]{usethis-package}} package to set up a new
R project in RSudio.
}
\details{
This function create a new project and does the following:
}
