% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umn_palette.R
\name{neyhart_palette}
\alias{neyhart_palette}
\title{Custom color palettes}
\usage{
neyhart_palette(
  name = c("umn1", "umn2", "umn3", "barley", "fall"),
  n,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{The name of the palette. See \emph{Details} for palette name options.}

\item{n}{The number of desired colors from the palette.}

\item{type}{The type of palette.}
}
\description{
Sample from some of my favorite color palettes
}
\details{
Palette options include:
\itemize{
  \item{"umn1"}{"University of Minnesota color scheme 1."}
  \item{"umn2"}{"University of Minnesota color scheme 2."}
  \item{"umn3"}{"University of Minnesota color scheme 3."}
  \item{"barley"}{"A color scheme from barley pictures."}
  \item{"fall"}{"An autumnal inspired color scheme"}
}

This function is inspired by \code{\link[wesanderson]{wes_palette}}.
}
