% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{neyhart_palettes}
\alias{neyhart_palettes}
\title{Custom color palettes}
\format{
A list of vectors containing hexadecimal color codes.
}
\source{
\url{https://www.ur.umn.edu/brand/assets/pdf/secondary_colors_rgb.pdf}
}
\usage{
neyhart_palettes
}
\description{
Hexadecimal codes to generate themes.
}
\keyword{datasets}
