% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_themes.R
\name{theme_poster}
\alias{theme_poster}
\alias{theme_presentation}
\alias{theme_presentation2}
\title{Plotting theme for presentations or posters}
\usage{
theme_poster(base_size = 16)

theme_presentation(base_size = 16)

theme_presentation2(base_size = 16)
}
\description{
Theme for ggplot2 designed for posters or presentations
}
\details{
\describe{
  \item{\code{theme_poster}}{Theme for posters.}
  \item{\code{theme_presentation}}{Theme for presentations.}
  \item{\code{theme_presentation2}}{Alternate theme for presentations.}
}
}
