% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{protocol}
\alias{protocol}
\alias{templ_protocol}
\title{Render a laboratory protocol document from the Neyhart template}
\usage{
protocol(...)

templ_protocol()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for protocols in the Neyhart Lab.
}
