% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{rapid_cv}
\alias{rapid_cv}
\title{Rapid cross-validation without removing samples}
\usage{
rapid_cv(object, index, rapid = TRUE)
}
\arguments{
\item{object}{A fitted model object}

\item{index}{A list of indices for leave-n-out cross-validation (i.e. the 
training set)}

\item{rapid}{Logical; should the rapid method be used?}
}
\description{
Implements cross-validation according to Gianola and Schon 2016. Returns
RMSE and r-squared
}
\examples{
fit <- lm(Sepal.Length ~ ., iris)

indices <- lapply(X = crossv_loo_grouped(group_by(iris, Species))$train, "[[", "idx")

}
