% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{rpospois}
\alias{rpospois}
\title{Sample from a positive Poisson distribution}
\usage{
rpospois(n, lambda)
}
\arguments{
\item{n}{The number of independent samples to draw from the distribution}

\item{lambda}{The mean parameter for a Poisson Distribution.}
}
\value{
A numeric vector with counts given a mean parameter \code{lambda}.
}
\description{
Generates random samples from a positive Poisson distribution. This distribution
ensures non-zero counts for every random sample.
}
\examples{
rpospois(n = 10, lambda = 1)

}
