% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_append.R
\name{save_append}
\alias{save_append}
\title{Append data to a RData file}
\usage{
save_append(..., list, file)
}
\arguments{
\item{...}{The names of objects to be saved.}

\item{list}{A character vector containing the names of objects to be saved.}

\item{file}{The name of a file where the data will be saved.}
}
\description{
Save one or more \code{R} objects to a RData file already containing objects.
}
