% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tassel2fastPHASE.R
\name{tassel2fastPHASE}
\alias{tassel2fastPHASE}
\title{Convert between TASSEL hapmap file and a fastPHASE input}
\usage{
tassel2fastPHASE(hapmap = NULL, filein = NULL, fileout)
}
\arguments{
\item{hapmap}{A TASSEL-encoded hapmap file read in to R as a data.frame. See \code{Details} 
for information on file format. Mutually exclusive with \code{filein}.}

\item{filein}{The path to a TASSEL-encoded hapmap file. See \code{Details} 
for information on file format. Mutually exclusive with \code{hapmap}.}

\item{file}{The output file name.}
}
\value{
Writes a fastPHASE-ready input file to the designated filepath.
}
\description{
Converts from a TASSEL hapmap file, which typically has diploid encoding of
genotypes (i.e. AA or CC or AC) to an input file for fastPHASE, which is formatted
line-by-line and contains information that is more haplotype-based.
}
\details{
The TASSEL format is as such:
The first row is column names. The first 4 columns are marker name, alleles, 
chromosome, and position, respectively. The next 7 column are additional information for 
TASSEL. The remaining columns are samples. Genotypes should be encoded in 
diploid format (i.e. AA, AC, CC) with "NN" denoting missing data.
}

