% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UM_colors.R
\name{UMN.color.palette}
\alias{UMN.color.palette}
\title{University of Minnesota Color Generator}
\usage{
UMN.color.palette(n)
}
\arguments{
\item{n}{Integer number of colors to be generated}
}
\value{
A character vector of hexidecimal color codes from the University
of Minnesota color palette.
}
\description{
Generates hexidecimal colors from the University of Minnesota
color palette.
}
\details{

}

