% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UM_colors.R
\docType{data}
\name{UMN.colors}
\alias{UMN.colors}
\title{University of Minnesota Color Palette}
\format{A character vector of hexidecimal codes.}
\source{
\url{https://www.ur.umn.edu/brand/assets/pdf/secondary_colors_rgb.pdf}
}
\usage{
UMN.colors
}
\description{
A vector of hexadecimal color codes that are designed to work well together with the University of Minnesota color scheme.
}
\keyword{datasets}

