% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umn_palette.R
\name{umn_palette}
\alias{umn_palette}
\title{University of Minnesota Color Generator}
\usage{
umn_palette(name = "Secondary_Tier1", n, type = c("discrete", "continous"))
}
\arguments{
\item{name}{The name of the secondary color tier. Options are "Secondary_Tier1",
"Secondary_Tier2", nad  "Secondary_Tier3".}

\item{n}{The number of desired colors from the palette.}

\item{type}{The type of palette.}
}
\value{
A \code{palette} vector of hexidecimal color codes from the University
of Minnesota color palette (see \code{\link[neyhart]{umn_palette}})
}
\description{
Deprecated.
}
\details{
This function is inspired by \code{\link[wesanderson]{wes_palette}}.
}
