% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umn_palette.R
\docType{data}
\name{umn_palettes}
\alias{umn_palettes}
\title{University of Minnesota Color Palette}
\format{
A nested list of character vectors of hexidecimal codes.
}
\source{
\url{https://www.ur.umn.edu/brand/assets/pdf/secondary_colors_rgb.pdf}
}
\usage{
umn_palettes
}
\description{
A vector of hexadecimal color codes that are designed to work 
well together with the University of Minnesota color scheme.
}
\keyword{datasets}
