% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_transformation.R
\name{ztrans}
\alias{ztrans}
\alias{zexp}
\title{Z-transformation}
\usage{
ztrans(x)

zexp(x)
}
\arguments{
\item{x}{A correlation coefficient estimate.}
}
\description{
Perform a z-transformation for modeling estimates for correlation coefficients and convert the z-score
back to a correlation coefficient.
}
