% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicore.R
\name{assign_cores}
\alias{assign_cores}
\title{Assign cores to a data.frame}
\usage{
assign_cores(df, n_core, split = FALSE)
}
\arguments{
\item{df}{A data.frame.}

\item{n_core}{The number of cores (integer) to split the data.frame.}

\item{split}{Logical. Should the data.frame be split by cores?}
}
\description{
Assigns the core (integer) to each row of a data.frame, given the desired number of cores.
}
