% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{confidInt}
\alias{confidInt}
\title{Mean, standard errors, and confidence intervals of samples}
\usage{
confidInt(x, level = 0.95, return.list = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{level}{The confidence level, or \eqn{1 - \alpha}.}

\item{return.list}{Logical. Should the output be a list (TRUE) or a vector (FALSE)?}

\item{mu}{A numer indicating the true value of the mean}

\item{alternative}{A character specifying the alternative hypothesis with
respect to \code{mu}.}
}
\description{
Mean, standard errors, and confidence intervals of samples
}
