% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\alias{new_analysis}
\alias{new_poster}
\alias{new_package}
\title{Create a new project}
\usage{
new_project(
  project.name,
  dir = ".",
  type = c("analysis", "poster", "package"),
  use.hpc = FALSE,
  use.git = TRUE,
  use.github = TRUE,
  use.pkgdown = FALSE
)

new_analysis(
  project.name,
  dir = ".",
  use.hpc = TRUE,
  use.git = TRUE,
  use.github = TRUE
)

new_poster(
  project.name,
  dir = ".",
  use.hpc = FALSE,
  use.git = TRUE,
  use.github = TRUE
)

new_package(
  project.name,
  dir = ".",
  use.pkgdown = TRUE,
  use.git = TRUE,
  use.github = TRUE
)
}
\arguments{
\item{project.name}{The name of the project.}

\item{dir}{The root directory of the new project.}

\item{type}{The type of project}

\item{use.hpc}{Logical. Should scripts be provided to run analysis on SCINet?}

\item{use.git}{Initialize a git repository?}

\item{use.github}{Connect to github?}
}
\description{
This function performs the setup for creating new projects of various types, 
including analyses, posters, and packages.
}
\details{
The function supports the creation of projects of the following types:

\describe{
  \item{analysis}{Directory setup and template scripts for conducting an analysis}
  \item{poster}{Directory setup and template copying for producing a poster in LaTeX}
  \item{package}{Setup for creating a new R package}
}
}
\section{Functions}{
\itemize{
\item \code{new_analysis()}: 

\item \code{new_poster()}: 

\item \code{new_package()}: 

}}
