% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_themes.R
\name{theme_neyhart}
\alias{theme_neyhart}
\alias{theme_acs}
\alias{theme_pnas}
\alias{theme_genetics}
\title{Plotting theme for publications}
\usage{
theme_neyhart(base_size = 8)

theme_acs(base_size = 8)

theme_pnas(base_size = 8)

theme_genetics(base_size = 8)
}
\description{
Themes for ggplot2 designed for compliance with variance scientific journals.
}
\details{
\describe{
  \item{\code{theme_acs}}{Theme for the American Society of Agronomy (ASA), 
  Crop Science Society of America (CSSA), and Soil Science Society of America (SSSA).}
  \item{\code{theme_pnas}}{Theme for the Proceedings of the National Academy of Sciences.}
  \item{\code{theme_genetics}}{Theme for GENETICS/G3.}
}
}
\examples{
g <- qplot(x = disp, y = hp, data = mtcars)


}
